Use protocol-relative links for admin-ajax calls
The plugin uses the admin_url() function to build all of its links to
ajax handlers. This fails under certain conditions: when a site uses
FORCE_SSL_ADMIN, these links are built using https:// as the protocol.

This can cause a number of cross-domain problems, especially in latest
FIrefox and in IE9+. If an https xd_reciever page is used on an http
page, then cross-domain communcation won't work properly. And if an
https url is used as the redirect_uri on an http page, that will
similarly fail in IE10.

The admin_url() function takes an optional second argument, $scheme, and
if an empty string is passed there, it will use the scheme of the
calling page, which is the behavior we want here.