Avert the Eye of Sauron's gaze
When using legacy (iframe) Capture signin, the logged-in cookies are
supposed to be set from the page specified as the redirect_uri, which
loads in the same iframe as the Capture widget. However, some
configurations of Internet Explorer (which were the default settings up
through IE8) refuse to set cookies from within an iframe, EVEN if the
iframe is on the same domain as the parent page, unless that iframe
sends a P3P (Platform for Privacy Preferences) header containing a
compact privacy policy (CP).

This adds a P3P header to the redirect_uri function. The policy I chose
is the same one that's hardcoded into the Capture widget... it may or
may not be legit, but it works to divert the Eye.

Source:
http://stackoverflow.com/questions/389456/cookie-blocked-not-saved-in-iframe-in-internet-explorer

Also, apply the same trick as in 7f07517 to force the logged-in cookie
to set relative to the protocol of the requesting page.